Add node location support
Currently we are running a Riak cluster in different locations (you can also call it site or availability zone).
The expectation is to be able to survive the loss of one entire site without any service interruption. The current  n_val is server_count-1, and the pw/pr is floor(n_val/2)+1. If we want to scale our Riak cluster for example double the cluster, we need to increase the n_val/pw/pr to meet the expectation.

Our idea is to be able to assemble a Riak ring where the neighbouring v_nodes are on different locations. This way we can reduce the replication factor to the number of locations + ~1 (to cover tail violations) which has the advantage that our data need less space, network load is reduced and we are able to scale our cluster without increasing the replication factor.

In this pr location has been introduced. When claiming a new ring the list of nodes is ordered taking into consideration the location of the individual nodes, in a manner that to adjacent nodes are preferably from different locations.